+++ tests/extras/test_rabbit.py
@@ -39,7 +39,7 @@ def test_connection_works_sibling_network(docker_client, vhost, create_and_pull)
             url = f"http://{aliases[0]}:{RABBIT_HTTP_API_PORT}/api/vhosts"
             container = create_and_pull(
                 docker_client,
-                "byrnedo/alpine-curl", f'-u guest:guest -vvv -I "{url}" --http0.9',
+                "byrnedo/alpine-curl:latest", f'-u guest:guest -vvv -I "{url}" --http0.9',
                 detach=True
             )
             with connect(network, container):
@@ -55,7 +55,7 @@ def test_connection_works_sibling(docker_client, host_ip, vhost, create_and_pull
         url = f"http://{host_ip}:{api_port}/api/vhosts"
         container = create_and_pull(
             docker_client,
-            "byrnedo/alpine-curl", f'-u guest:guest -vvv -I "{url}" --http0.9',
+            "byrnedo/alpine-curl:latest", f'-u guest:guest -vvv -I "{url}" --http0.9',
             detach=True
         )
         container.start()