+++ yellowbox/extras/redis.py
@@ -6,7 +6,7 @@ from redis import ConnectionError as RedisConnectionError, Redis
 
 from yellowbox.containers import get_ports, create_and_pull, upload_file
 from yellowbox.subclasses import SingleContainerService, RunMixin
-from yellowbox.utils import retry
+from yellowbox.utils import RetryMixin
 
 __all__ = ['RedisService', 'REDIS_DEFAULT_PORT', 'DEFAULT_RDB_PATH', 'append_state']
 
@@ -28,7 +28,7 @@ def append_state(client: Redis, db_state: RedisState):
             client.set(k, v)
 
 
-class RedisService(SingleContainerService, RunMixin):
+class RedisService(RetryMixin, SingleContainerService, RunMixin):
     def __init__(self, docker_client: DockerClient, image='redis:latest',
                  redis_file: Optional[IO[bytes]] = None, **kwargs):
         container = create_and_pull(docker_client, image, publish_all_ports=True, detach=True)
@@ -53,7 +53,7 @@ class RedisService(SingleContainerService, RunMixin):
     def start(self):
         super().start()
         with self.client() as client:
-            retry(client.ping, RedisConnectionError)
+            self.retry(client.ping, RedisConnectionError)
         self.started = True
         return self